rename: posts-manager 폴더 생성
