+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { BrowserRouter as Router } from "react-router-dom"
 import Header from "./widgets/header/ui/Header.tsx"
 import Footer from "./widgets/footer/ui/Footer.tsx"
-import PostsManagerPage from "./pages/ui/PostsManagerPage.tsx"
+import PostsManagerPage from "./pages/posts-manager/ui/PostsManagerPage.tsx"
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query"
 
 const queryClient = new QueryClient({
